[#347][IMP] frontend_listing: add list.
