+++ main/local_modules/frontend/templates/website_partner.xml
@@ -1,16 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <openerp>
     <data>
-        <!--Disbaled to be able to release code online for now.-->
-        <record id="directory" model="website.menu">
-            <field name="name">Directory</field>
-            <field name="url">/directory</field>
-            <field name="parent_id" ref="website.main_menu"/>
-            <field name="user_logged" eval="False"/>
-            <field name="user_not_logged" eval="False"/>
-            <field name="sequence" type="int">20</field>
-        </record>
-
         <template id="frontend.partner" name="Frontend: partner" page="True">
             <t t-call="website.layout">
                 <div class="modal-content" style="margin:30px">
