+++ main/local_modules/frontend_listing/__openerp__.py
@@ -36,6 +36,7 @@
         'templates/template_engine.xml',
         'templates/template_head.xml',
         # 'templates/template_html.xml',
+        'data/website_menus.xml',
     ],
     'installable': True,
 }
