+++ main/local_modules/frontend_listing/controllers/__init__.py
@@ -19,4 +19,4 @@
 #
 ##############################################################################
 
-from . import homepage
+from . import listing
