+++ main/local_modules/frontend_listing/templates/template_body.xml
@@ -1,14 +1,22 @@
 <?xml version="1.0" encoding="utf-8"?>
 <openerp>
     <data>
-        <template id="frontend_listing.search" name="Search hidden fields">
-            <form t-att-action="keep('/directory', search=0)" method="get" t-att-class="search_class">
-
+        <!--Requirements:-->
+        <!--* url-->
+        <template id="frontend_listing.pagination" name="Listing: Pagination">
+            <form t-att-action="keep(url, search=0)" method="get" t-att-class="search_class">
                 <t t-if="attrib_values">
                     <t t-foreach="attrib_values" t-as="a">
                         <input type="hidden" name="attrib" t-att-value="'%s-%s' % (a[0], a[1])"/>
                     </t>
                 </t>
+            </form>
+        </template>
+
+        <!--Requirements:-->
+        <!--* url-->
+        <template id="frontend_listing.search" name="Listing: search">
+            <form t-att-action="keep(url, search=0)" method="get" t-att-class="search_class">
                 <div class="input-group">
                     <input type="text" name="search" class="search-query form-control" placeholder="Search..." t-att-value="search"/>
                     <span class="input-group-btn">
@@ -17,16 +25,23 @@
                 </div>
 
             </form>
-            <a href="/web/signup">Sign up for more options. It is free!</a>
-        </template>
+         </template>
 
-        <template id="frontend_listing.listing" name="Base for listing of companies">
+        <template id="frontend_listing.base" name="Base for listing of companies">
             <t t-call="frontend_base.doctype_html"/>
             <t t-call="frontend_base.html">
 
                 <t t-call="frontend_base.full_head"/>
 
                 <t t-call="frontend_base.body">
+                    <div class="products_pager">
+                        <div class="row" style="width: 100%;">
+                            <t t-call="frontend_listing.search">
+                                <t t-set="search_class">pagination form-inline col-md-3</t>
+                            </t>
+                        </div>
+                    </div>
+
                     <t t-raw="0"/>
                     <t t-call="frontend_base.footer"/>
                     <t t-call="frontend_base.engine"/>
@@ -37,11 +52,46 @@
 
 
         <template id="frontend_listing.map" name="Listing: Map">
-            <t t-call="frontend_listing.listing">
-                <div id="map">MAP</div>
+            <t t-call="frontend_listing.base">
+                <div id="map"></div>
             </t>
             <script type="text/javascript" src="/frontend_listing/static/src/js/google_map.js"></script>
         </template>
 
+        <template id="frontend_listing.list" name="Listing: List">
+            <t t-call="frontend_listing.base">
+               <table data-toggle="table">
+                    <thead>
+                        <tr>
+                            <th/>
+                            <th>Name</th>
+                            <th>City</th>
+                            <th>Country</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        <t t-foreach="partners" t-as="partner">
+                        <tr>
+                            <td>
+                                <img itemprop="image" class="img img-responsive" t-att-src="website.image_url(partner, 'image_small')" />
+                            </td>
+                            <td>
+                            <a itemprop="name" t-att-href="keep('/directory/company/%s' % slug(partner), page=(pager['page']['num'] if pager['page']['num']>1 else None))">
+                                <h4 t-esc="partner.name"/>
+                            </a>
+                            </td>
+                            <td><span t-esc="partner.city"/></td>
+                            <td><span t-esc="partner.country_id.name"/></td>
+                        </tr>
+                        </t>
+                    </tbody>
+                </table>
+
+                <t t-call="frontend_listing.pagination">
+                    <t t-set="search_class">pagination form-inline col-md-3</t>
+                </t>
+                <t t-call="website.pager"/>
+            </t>
+        </template>
     </data>
 </openerp>
